requirements.txt
"""
Required packages for Banking Fraud Detection API
"""

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
PyJWT==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.25.2
pandas==2.1.4
cryptography==41.0.7
passlib[bcrypt]==1.7.4
pyotp==2.9.0
